Updating branch + adding Mobile Messaging support

+++ backend/app/routes/users.py
@@ -1,10 +1,11 @@
-from app.rbac_utilities import create_mysql_user, safe_db_connection, is_user_authorized, is_user_authenticated
+from app.rbac_utilities import create_mysql_user, safe_db_connection, is_user_authorized, is_user_authenticated, update_mysql_user
 from app.extensions import db, bcrypt
-from flask import Blueprint, jsonify, request, session, make_response
+from flask import Blueprint, jsonify, request, session, make_response, current_app
 from app.models.user import (User, Listing, Chat, Message, user_chat_association,
                              user_favorited_listing_association, user_passed_listing_association,
                              MAX_FIRST_NAME_LENGTH, MAX_LAST_NAME_LENGTH, MAX_PASSWORD_LENGTH, MAX_EMAIL_LENGTH,
-                             MAX_USERNAME_LENGTH)
+                             MAX_USERNAME_LENGTH, Address, Image)
+import os
 
 user_bp = Blueprint('user', __name__)
 
@@ -23,9 +24,38 @@ def load_moderator_page():
     authorization = is_user_authorized('moderator')
     if isinstance(authorization, tuple):
         return authorization
+    
+    with safe_db_connection(session.get('username'), session.get('password')) as moderator_db_session:
+        all_listings = moderator_db_session.query(Listing).all()
+        all_listings_data = [listing.to_dict() for listing in all_listings]
+        return jsonify(all_listings_data), 200
 
-    return jsonify({'message': 'Authorized'})
+@user_bp.route('/moderator/delete', methods=['DELETE'])
+def delete_a_listing_as_moderator():
+    authorization = is_user_authorized('moderator')
+    if isinstance(authorization, tuple):
+        return authorization
+    
+    data = request.get_json()
+    listing_id = data.get('listing_id')
 
+    if not listing_id:
+        return jsonify({'error': 'Listing ID is required.'});
+
+    with safe_db_connection(session.get('username'), session.get('password')) as moderator_db_session:
+            moderator_db_session.query(Address).filter(Address.listing_id == listing_id).delete(synchronize_session='fetch')
+            moderator_db_session.query(Image).filter(Image.listing_id == listing_id).delete(synchronize_session='fetch')
+            moderator_db_session.query(user_favorited_listing_association).filter(user_favorited_listing_association.c.listing_id == listing_id).delete(synchronize_session='fetch')
+            moderator_db_session.query(user_passed_listing_association).filter(user_passed_listing_association.c.listing_id == listing_id).delete(synchronize_session='fetch')
+            listing_to_delete = moderator_db_session.query(Listing).get(listing_id)
+            
+            if listing_to_delete is None:
+                return jsonify({'error': 'Listing not found.'}), 404
+
+            moderator_db_session.delete(listing_to_delete)
+            moderator_db_session.commit()
+
+            return jsonify({'message': 'Listing deleted successfully.'}), 200
 
 @user_bp.route('profile', methods=['GET'])
 def get_current_user():
@@ -38,7 +68,7 @@ def get_current_user():
     with safe_db_connection(session.get('username'), session.get('password')) as user_db_session:
         user = user_db_session.query(User).filter_by(id=user_id).first()
 
-    return jsonify({'id': user.id, 'username': user.username, 'first_name': user.first_name, 'last_name': user.last_name})
+    return jsonify({'id': user.id, 'username': user.username, 'first_name': user.first_name, 'last_name': user.last_name, 'email': user.email})
 
 
 '''
@@ -139,35 +169,84 @@ def register_user():
     return jsonify({'error': 'Missing required fields'}), 400
 
 
-'''
-IMPORTANT:
-THIS HEADER DENOTES THAT THE FOLLOWING API ROUTE MEETS ONE OF THE FOLLOWING CRITERIA:
-- API ROUTE IS NEVER USED IN THE CLIENT-SIDE APPLICATION
-- API ROUTE NEEDS RBAC IMPLEMENTED IN IT IF NECESSARY (A USER DB CONNECTION PERFORMING ACTIONS ON THEIR BEHALF, NOT THE ADMIN DB CONNECTION)
-'''
+@user_bp.route('/editprofile', methods=['PUT'])
+def update_user():
+    authorization = is_user_authorized('user')
+    if isinstance(authorization, tuple):
+        return authorization
 
-
-@user_bp.route('<int:user_id>', methods=['PUT'])
-def update_user(user_id):
     user_json = request.get_json()
-    user = User.query.get(user_id)
 
-    updatable_fields = ['first_name', 'last_name',
-                        'email', 'username', 'password']
+    required_fields = ['first_name', 'last_name', 'username', 'email']
+    if not all(field in user_json for field in required_fields):
+        return jsonify({'error': 'Missing required fields'}), 400
 
-    if not user:
-        return jsonify({'error': 'User not found'}), 404
+    current_user_id = session.get('user_id')
 
-    if not any(field in user_json for field in updatable_fields):
-        return jsonify({'message': 'At least one updatable field must be provided'})
+    with safe_db_connection(session.get('username'), session.get('password')) as user_db_session:
+        if user_db_session.query(User).filter(User.username == user_json['username'], User.id != current_user_id).first():
+            return jsonify({'error': 'A user already exists with the provided username'}), 409
 
-    for field in updatable_fields:
-        if field in user_json:
-            setattr(user, field, user_json[field])
+        if (len(user_json['first_name']) > MAX_FIRST_NAME_LENGTH or
+            len(user_json['last_name']) > MAX_LAST_NAME_LENGTH or
+                len(user_json['username']) > MAX_USERNAME_LENGTH):
+            return jsonify({'error': 'One or more input fields are over the maximum character limit', 'code': 'MAX_INPUT_LIMIT'}), 400
 
-    db.session.commit()
+        user = user_db_session.query(User).filter_by(
+            id=current_user_id).first()
+        if user:
+            if len(user_json['first_name']) != 0:
+                user.first_name = user_json['first_name']
+            if len(user_json['last_name']) != 0:
+                user.last_name = user_json['last_name']
+            if len(user_json['username']) != 0:
+                user.username = user_json['username']
+                update_mysql_user(user.username)
+            if len(user_json['email']) != 0:
+                user.email = user_json['email']
 
-    return jsonify({'message': 'User updated successfully', **user.to_dict()}), 200
+            user_db_session.commit()
+            return jsonify({'message': 'User updated successfully'}), 200
+        else:
+            return jsonify({'error': 'User not found'}), 404
+
+
+@user_bp.route('/resetpassword', methods=['PUT'])
+def reset_password():
+    authorization = is_user_authorized('user')
+    if isinstance(authorization, tuple):
+        return authorization
+
+    user_json = request.get_json()
+
+    user_id = session.get('user_id')
+
+    if 'password' not in user_json or not user_json['password']:
+        return jsonify({'error': 'Password field is required'}), 400
+
+    if len(user_json['password']) > MAX_PASSWORD_LENGTH:
+        return jsonify({'error': 'Password exceeds maximum length allowed', 'code': 'MAX_INPUT_LIMIT'}), 400
+
+    current_user_id = session.get('user_id')
+    with safe_db_connection(session.get('username'), session.get('password')) as user_db_session:
+        user = user_db_session.query(User).filter_by(
+            id=current_user_id).first()
+
+        if user:
+            print(user_json['password'])
+            user_json['password'] = bcrypt.generate_password_hash(
+                user_json['password'])
+
+            user.password = user_json['password']
+            user_db_session.commit()
+
+            user = user_db_session.query(User).filter_by(id=user_id).first()
+
+            update_mysql_user(user.username, user.password)
+
+            return jsonify({'message': 'Password reset successfully'}), 200
+        else:
+            return jsonify({'error': 'User not found'}), 404
 
 
 '''
@@ -178,14 +257,23 @@ THIS HEADER DENOTES THAT THE FOLLOWING API ROUTE MEETS ONE OF THE FOLLOWING CRIT
 '''
 
 
-@user_bp.route('<int:user_id>', methods=['DELETE'])
-def delete_user(user_id):
+@user_bp.route('delete', methods=['DELETE'])
+def delete_user():
+    authorization = is_user_authorized('user')
+    if isinstance(authorization, tuple):
+        return authorization
+
+    user_id = session.get('user_id') or request.args.get('user_id')
+
+    if not user_id:
+        return jsonify({'error': 'No user ID provided'}), 400
+
     user = User.query.get(user_id)
 
     if user:
         db.session.delete(user)
         db.session.commit()
-
+        session.pop('user_id', None)
         return jsonify({'message': 'User deleted successfully'}), 200
     else:
         return jsonify({'error': 'User not found'}), 404
@@ -342,7 +430,6 @@ def favorite_a_listing_for_user():
             user_db_session.add(chat)
             user_db_session.commit()
 
-            # Create associations for both users
             new_association = {'user_id': user_id, 'chat_id': chat.id}
             user_db_session.execute(
                 user_chat_association.insert().values(new_association))
@@ -387,30 +474,32 @@ THIS HEADER DENOTES THAT THE FOLLOWING API ROUTE MEETS ONE OF THE FOLLOWING CRIT
 '''
 
 
-@user_bp.route('<int:user_id>/favorite-listings/<int:listing_id>', methods=['DELETE'])
-def unfavorite_a_listing_from_user(user_id, listing_id):
-    user = User.query.get(user_id)
-    listing = Listing.query.get(listing_id)
+@user_bp.route('favorite-listings', methods=['DELETE'])
+def unfavorite_a_listing_from_user():
+    authorization = is_user_authorized('user')
+    if isinstance(authorization, tuple):
+        return authorization
+    
+    user_id = session.get('user_id')
 
-    if not user:
-        return jsonify({'error': 'User not found'}), 404
+    listing_id_json = request.get_json()
+    if 'listing_id' not in listing_id_json:
+        return jsonify({'error': 'Missing listing id'}), 404
+    
+    listing_id = listing_id_json['listing_id']
 
-    if not listing:
-        return jsonify({'error': 'Listing not found'}), 404
+    with safe_db_connection(session.get('username'), session.get('password')) as user_db_session:
+        listing = Listing.query.get(listing_id)
+        if not listing:
+            return jsonify({'error': 'Listing not found'}), 404
+        
+        association = user_favorited_listing_association.delete().where(
+            (user_favorited_listing_association.c.user_id == user_id) &
+            (user_favorited_listing_association.c.listing_id == listing_id)
+        )
 
-    association_exists = db.session.query(db.exists().where(
-        (user_favorited_listing_association.c.user_id == user_id) &
-        (user_favorited_listing_association.c.listing_id == listing_id)
-    )).scalar()
-
-    if not association_exists:
-        return jsonify({'error': 'Cannot unfavorite the specified listing from the user as it was not favorited'}), 404
-
-    db.session.execute(user_favorited_listing_association.delete().where(
-        (user_favorited_listing_association.c.user_id == user_id) &
-        (user_favorited_listing_association.c.listing_id == listing_id)
-    ))
-    db.session.commit()
+        db.session.execute(association) 
+        db.session.commit()
 
     return jsonify({'message': 'Listing successfully unfavorited from the user'}), 200
